ft_atoi 리턴 값 (int)로 캐스팅 C에서 큰 자료형에서 작은 자료형으로 바꿀 때 그냥 암시적으로 됌 그럼에도 평가를 위해서 명시적으로 (int)로 변경해줌

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jokang <autoba9687@gmail.com>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 15:14:33 by jokang            #+#    #+#             */
-/*   Updated: 2021/11/27 15:40:16 by jokang           ###   ########.fr       */
+/*   Updated: 2021/12/07 17:55:44 by jokang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,5 +47,5 @@ int	ft_atoi(const char *str)
 			return (0);
 		str++;
 	}
-	return (sign * result);
+	return ((int)(sign * result));
 }
